requirements.txt
appdirs==1.4.3
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==3.0.4
django-import-export==2.1.0
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
psycopg2==2.8.5
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
virtualenv==20.0.16
virtualenvwrapper-win==1.2.6
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
